requirements.txt
aiogram==3.2.0
sqlalchemy==2.0.23
python-dotenv==1.0.0
yookassa==3.0.0
